initialise cargo workspace, add first project
using a [cargo workspace][] to manage the drivers in separate projects
but still within the same repository for now. in a future step, when
they've matured a bit, they could possibly be moved to own repos and be
published.

the initial project was generated using `cargo new`, no changes have
been done to it yet.

[cargo workspace]: https://doc.rust-lang.org/book/ch14-03-cargo-workspaces.html