add initial emergency brake option
this doesn't work 100% as intended yet due to the way the logic for the
forward/backward speed of the remote control has been implemented: the
remote control currently stores the speed it thinks the motor has and
uses that on the next button press. however, if we do an emergency
brake, the car stops the motor but the car doesn't know about the remote
control and thus can't let it know the new speed. accordingly, the next
time the speed is updated via the remote control it'll have a wrong
value (because it doesn't think that the car is stopped).

this will be fixed in a follow-up commit by changing the remote control
to query the speed from the car instead of trying to store it by itself.