implement the display
this requires using `shared-bus` to share the I2C bus between the
devices connected to it (so far only one device was connected, so it was
ok to just consume the I2C there directly; now this is no longer ok).

in a first step nothing sensible is written out yet, this will be added
in a second step because it needs a bit more work (most likely an
allocator will be needed to properly format the message).