make all I2C devices optional
the IMU (which isn't actually used in the code at this point) seems to
have caused issues as suddenly no I2C communication worked anymore while
it was connected.
to debug this, the TOF and display (both of which are using I2C) have
been made optional in the code (neither is strictly necessary for a pure
RC car). they are enabled by default, however by compiling without
default features they can be removed.

this has been done in a way which didn't require a lot of code changes,
thus the dependencies for the TOF & display are still being compiled
into the binary even if they're not enabled/used. in the best case all
references to them would've been #[cfg] guarded and the dependencies
made optional (enabled with the corresponding features).
but since the standard operation mode is with these features, time is
scarce and this project probably won't be used for long beyond the
finishing of the course, this route was taken...

the final fix was to remove the IMU (MPU6050) - there might be a
hardware fault with it (unclear, not further tested as it's currently
unused).