add more detailed logging on boot
this is intentionally done using `info` rather than `debug` so that it's
always available (as it's only on startup it's not spamming a lot). this
way, the information is readily available if something doesn't work.

this is similar to how e.g. Linux prints a lot of messages on startup
(and even for Windows it's possible to switch off the boot logo and see
the startup messages instead), which helps if the system fails to start
up to narrow down which step went wrong.