clear BT RX buffer after processing
the way the buffer is used in the DMA operation means that it's never
cleared out. as long as the new messages are the same length (or longer)
than the previous ones means that they just overwrite each other and
everything is fine.

but if the previous message was longer (which can happen if multiple
button presses happen fast enough one after the other, thus leading to
no line idle interrupt in-between) the next message will only overwrite
the first command(s) and the other old commands will stay in there and
be processed again.

this led to ugly bugs where you'd e.g. press the stop button but the car
would immediatly start driving again because before you e.g. pressed 2x
up fast, so the last two commands (press + release of the up button)
would stay in there.

fixes #28